feat(orchestrator): refactor ecp to orchestrator
