+++ .github/workflows/ci-it.yml
@@ -160,7 +160,7 @@ jobs:
           - ./internal/tools/monitor/...
           - ./internal/core/...
           - ./internal/apps/dop/...
-          - ./internal/apps/admin/... ./internal/apps/devflow/... ./internal/apps/ecp/... ./internal/apps/gallery/...
+          - ./internal/apps/admin/... ./internal/apps/devflow/... ./internal/apps/gallery/...
           - ./internal/apps/msp/...
           - ./internal/apps/cmp/...
     steps:
