+++ bundle/ecp.go
@@ -49,7 +49,7 @@ func (b *Bundle) ListEdgeApp(req *apistructs.EdgeAppListPageRequest, identify ap
 		reqParam   map[string]string
 	)
 
-	host, err := b.urls.ECP()
+	host, err := b.urls.Orchestrator()
 	if err != nil {
 		return nil, err
 	}
@@ -118,7 +118,7 @@ func (b *Bundle) GetEdgeApp(appID uint64, identify apistructs.Identity) (*apistr
 		httpReqRes httpserver.Resp
 	)
 
-	host, err := b.urls.ECP()
+	host, err := b.urls.Orchestrator()
 	if err != nil {
 		return nil, err
 	}
@@ -162,7 +162,7 @@ func (b *Bundle) CreateEdgeApp(req *apistructs.EdgeAppCreateRequest, identify ap
 		resp httpserver.Resp
 	)
 
-	host, err := b.urls.ECP()
+	host, err := b.urls.Orchestrator()
 	if err != nil {
 		return err
 	}
@@ -191,7 +191,7 @@ func (b *Bundle) UpdateEdgeApp(req *apistructs.EdgeAppUpdateRequest, appID uint6
 		resp httpserver.Resp
 	)
 
-	host, err := b.urls.ECP()
+	host, err := b.urls.Orchestrator()
 	if err != nil {
 		return err
 	}
@@ -220,7 +220,7 @@ func (b *Bundle) DeleteEdgeApp(appID int64, identify apistructs.Identity) error 
 		resp httpserver.Resp
 	)
 
-	host, err := b.urls.ECP()
+	host, err := b.urls.Orchestrator()
 	if err != nil {
 		return err
 	}
@@ -250,7 +250,7 @@ func (b *Bundle) GetEdgeAppStatus(req *apistructs.EdgeAppStatusListRequest, iden
 		httpReqRes httpserver.Resp
 	)
 
-	host, err := b.urls.ECP()
+	host, err := b.urls.Orchestrator()
 	if err != nil {
 		return nil, err
 	}
@@ -306,7 +306,7 @@ func (b *Bundle) ListEdgeSite(req *apistructs.EdgeSiteListPageRequest, identify 
 		reqParam   map[string]string
 	)
 
-	host, err := b.urls.ECP()
+	host, err := b.urls.Orchestrator()
 	if err != nil {
 		return nil, err
 	}
@@ -384,7 +384,7 @@ func (b *Bundle) GetEdgeSite(siteID int64, identify apistructs.Identity) (*apist
 		httpReqRes httpserver.Resp
 	)
 
-	host, err := b.urls.ECP()
+	host, err := b.urls.Orchestrator()
 	if err != nil {
 		return nil, err
 	}
@@ -429,7 +429,7 @@ func (b *Bundle) CreateEdgeSite(req *apistructs.EdgeSiteCreateRequest, identify 
 		resp httpserver.Resp
 	)
 
-	host, err := b.urls.ECP()
+	host, err := b.urls.Orchestrator()
 	if err != nil {
 		return err
 	}
@@ -458,7 +458,7 @@ func (b *Bundle) UpdateEdgeSite(req *apistructs.EdgeSiteUpdateRequest, siteID in
 		resp httpserver.Resp
 	)
 
-	host, err := b.urls.ECP()
+	host, err := b.urls.Orchestrator()
 	if err != nil {
 		return err
 	}
@@ -487,7 +487,7 @@ func (b *Bundle) DeleteEdgeSite(siteID int64, identify apistructs.Identity) erro
 		resp httpserver.Resp
 	)
 
-	host, err := b.urls.ECP()
+	host, err := b.urls.Orchestrator()
 	if err != nil {
 		return err
 	}
@@ -518,7 +518,7 @@ func (b *Bundle) ListEdgeConfigset(req *apistructs.EdgeConfigSetListPageRequest,
 		reqParam   map[string]string
 	)
 
-	host, err := b.urls.ECP()
+	host, err := b.urls.Orchestrator()
 	if err != nil {
 		return nil, err
 	}
@@ -591,7 +591,7 @@ func (b *Bundle) GetEdgeConfigSet(itemID int64, identify apistructs.Identity) (*
 		httpReqRes httpserver.Resp
 	)
 
-	host, err := b.urls.ECP()
+	host, err := b.urls.Orchestrator()
 	if err != nil {
 		return nil, err
 	}
@@ -635,7 +635,7 @@ func (b *Bundle) CreateEdgeConfigset(req *apistructs.EdgeConfigSetCreateRequest,
 		resp httpserver.Resp
 	)
 
-	host, err := b.urls.ECP()
+	host, err := b.urls.Orchestrator()
 	if err != nil {
 		return err
 	}
@@ -664,7 +664,7 @@ func (b *Bundle) UpdateEdgeConfigset(req *apistructs.EdgeConfigSetUpdateRequest,
 		resp httpserver.Resp
 	)
 
-	host, err := b.urls.ECP()
+	host, err := b.urls.Orchestrator()
 	if err != nil {
 		return err
 	}
@@ -693,7 +693,7 @@ func (b *Bundle) DeleteEdgeConfigset(siteID int64, identify apistructs.Identity)
 		resp httpserver.Resp
 	)
 
-	host, err := b.urls.ECP()
+	host, err := b.urls.Orchestrator()
 	if err != nil {
 		return err
 	}
@@ -723,7 +723,7 @@ func (b *Bundle) ListEdgeCfgSetItem(req *apistructs.EdgeCfgSetItemListPageReques
 		buffer     bytes.Buffer
 	)
 
-	host, err := b.urls.ECP()
+	host, err := b.urls.Orchestrator()
 	if err != nil {
 		return nil, err
 	}
@@ -785,7 +785,7 @@ func (b *Bundle) GetEdgeCfgSetItem(itemID int64, identify apistructs.Identity) (
 		httpReqRes httpserver.Resp
 	)
 
-	host, err := b.urls.ECP()
+	host, err := b.urls.Orchestrator()
 	if err != nil {
 		return nil, err
 	}
@@ -830,7 +830,7 @@ func (b *Bundle) CreateEdgeCfgSetItem(req *apistructs.EdgeCfgSetItemCreateReques
 		resp httpserver.Resp
 	)
 
-	host, err := b.urls.ECP()
+	host, err := b.urls.Orchestrator()
 	if err != nil {
 		return err
 	}
@@ -859,7 +859,7 @@ func (b *Bundle) UpdateEdgeCfgSetItem(req *apistructs.EdgeCfgSetItemUpdateReques
 		resp httpserver.Resp
 	)
 
-	host, err := b.urls.ECP()
+	host, err := b.urls.Orchestrator()
 	if err != nil {
 		return err
 	}
@@ -888,7 +888,7 @@ func (b *Bundle) DeleteEdgeCfgSetItem(siteID int64, identify apistructs.Identity
 		resp httpserver.Resp
 	)
 
-	host, err := b.urls.ECP()
+	host, err := b.urls.Orchestrator()
 	if err != nil {
 		return err
 	}
@@ -1082,7 +1082,7 @@ func (b *Bundle) GetEdgeSiteInitShell(siteID int64, identify apistructs.Identity
 		buffer     bytes.Buffer
 	)
 
-	host, err := b.urls.ECP()
+	host, err := b.urls.Orchestrator()
 	if err != nil {
 		return nil, err
 	}
@@ -1215,7 +1215,7 @@ func (b *Bundle) RestartEdgeAppSite(appID uint64, req *apistructs.EdgeAppSiteReq
 		resp httpserver.Resp
 	)
 
-	host, err := b.urls.ECP()
+	host, err := b.urls.Orchestrator()
 	if err != nil {
 		return err
 	}
@@ -1244,7 +1244,7 @@ func (b *Bundle) OfflineEdgeAppSite(appID uint64, req *apistructs.EdgeAppSiteReq
 		resp httpserver.Resp
 	)
 
-	host, err := b.urls.ECP()
+	host, err := b.urls.Orchestrator()
 	if err != nil {
 		return err
 	}
