--- cmd/ecp/common-conf
@@ -1,1 +1,0 @@
-../../conf
