+++ erda.yml
@@ -501,25 +501,6 @@ services:
         path: "/health"
         duration: 120
       exec: { }
-  ecp:
-    cmd: /erda/cmd/ecp/bin
-    ports:
-      - port: 9029
-        protocol: "TCP"
-        l4_protocol: "TCP"
-    resources:
-      cpu: ${request_cpu:0.2}
-      max_cpu: 0.2
-      mem: 128
-      network:
-        mode: "container"
-    deployments:
-      replicas: 1
-      labels:
-        GROUP: "dice"
-    health_check:
-      http: { }
-      exec: { }
   cmp:
     cmd: /erda/cmd/cmp/bin
     envs:
