+++ internal/tools/orchestrator/ecp/endpoints/edge_site.go
@@ -24,7 +24,7 @@ import (
 	"github.com/sirupsen/logrus"
 
 	"github.com/erda-project/erda/apistructs"
-	"github.com/erda-project/erda/internal/apps/ecp/services/apierrors"
+	"github.com/erda-project/erda/internal/tools/orchestrator/ecp/services/apierrors"
 	"github.com/erda-project/erda/pkg/http/httpserver"
 	"github.com/erda-project/erda/pkg/strutil"
 )
