+++ internal/tools/orchestrator/ecp/services/edge/configset.go
@@ -22,7 +22,7 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/erda-project/erda/apistructs"
-	"github.com/erda-project/erda/internal/apps/ecp/dbclient"
+	"github.com/erda-project/erda/internal/tools/orchestrator/ecp/dbclient"
 )
 
 var (
